Moved Controllers to their own package. Created Services package with a PostService with 2 methods
