+++ src/main/java/com/codeup/springbootblog/Controllers/DiceController.java
@@ -1,4 +1,4 @@
-package com.codeup.springbootblog;
+package com.codeup.springbootblog.Controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
